Move aws_h1_stream to its own file
